Move ExecutableQuery out of ExecutionUtil.
