+++ test/integration/tpcc_execution_test.cpp
@@ -1,8 +1,10 @@
 #include "execution/exec/execution_context.h"
 #include "execution/exec/output.h"
-#include "execution/execution_util.h"
+#include "execution/executable_query.h"
+#include "execution/vm/module.h"
 #include "main/db_main.h"
 #include "parser/postgresparser.h"
+#include "planner/plannodes/abstract_plan_node.h"
 #include "test_util/test_harness.h"
 #include "test_util/tpcc/builder.h"
 #include "test_util/tpcc/database.h"
